+++ tests/testthat/test-Lesson.R
@@ -10,14 +10,68 @@ test_that("Lesson class will fail if given a bad path", {
 
 })
 
+test_that("styles-based lessons can not read in built files", {
+  expect_message(frg$load_built(), "Only lessons using sandpaper can load built files")
+})
+
+
 test_that("Sandpaper lessons can be read", {
   snd <- Lesson$new(path = lesson_fragment("sandpaper-fragment"), jekyll = FALSE)
   expect_s3_class(snd, "Lesson")
   expect_named(snd$episodes, "intro.Rmd")
+  expect_s3_class(snd$episodes[[1]], "Episode")
+  expect_null(snd$built)
+  expect_s3_class(snd$extra[[1]], "Episode")
+})
+
+test_that("Sandpaper lessons have getter and summary methods", {
+  snd <- Lesson$new(path = lesson_fragment("sandpaper-fragment"), jekyll = FALSE)
   # sandpaper lessons will have their divs pre-labeled.
   expect_length(snd$challenges()[[1]], 1L)
   expect_length(snd$solutions()[[1]], 2L)
+  expect_null(snd$get())
+  expect_length(snd$get("headings")[[1]], 6L)
+  expect_length(snd$get("code", TRUE)[[1]], 3L)
+  expect_length(snd$get("links")[[1]], 1L)
+  expect_length(snd$get("images")[[1]], 0L)
+  withr::local_options(list(width = 200))
+  # summary for all files can exist
+  expect_snapshot(snd$summary(TRUE))
+  # summary for episodes can exist
+  expect_snapshot(snd$summary())
 })
+
+
+test_that("Sandpaper lessons can read in built files", {
+
+  snd <- Lesson$new(path = lesson_fragment("sandpaper-fragment"), jekyll = FALSE)
+  snd$load_built()
+  withr::local_options(list(width = 200))
+  expect_snapshot(snd$summary(TRUE))
+  expect_snapshot(snd$summary("built"))
+
+})
+
+
+test_that("Sandpaper lessons will throw a warning for $load_built()", {
+  lf <- lesson_fragment("sandpaper-fragment")
+  withr::defer({
+    fs::file_move(
+      fs::path(lf, "site", "tliub"),
+      fs::path(lf, "site", "built")
+    )
+  })
+  fs::file_move(
+    fs::path(lf, "site", "built"),
+    fs::path(lf, "site", "tliub")
+  )
+  snd <- Lesson$new(path = lf, jekyll = FALSE)
+  expect_message(snd$load_built(),
+    "No files built. Run `sandpaper\\:\\:build_lesson\\(\\)` to build.")
+  expect_null(snd$built)
+})
+
+
 
 test_that("Sandpaper lessons can create handouts", {