+++ NEWS.md
@@ -1,4 +1,36 @@
-# pegboard 0.2.7
+# pegboard 0.3.0
+
+## NEW FEATURES
+
+### Episode class objects
+
+ - `$summary()` method which can summarise counts of elements in the episode.
+ - fixes for `$error` and `$output` active bindings
+ - new `$warning` active binding that will show code blocks with the `warning`
+   class.
+
+### Lesson class objects
+
+ - new public field "built" that will contain XML representations of 
+   markdown files built from RMarkdown files in sandpaper lessons.
+ - new public field "sandpaper" is a boolean that indicates if a lesson can be
+   built with sandpaper.
+ - new `$load_built()` method will load the built files if they exist in a
+   sandpaper lesson.
+ - new `$get()` method which will get any element from any Episode class object
+   contained within.
+ - new `$summary()` method which will call the `$summary()` method for any
+   Episode class object.
+
+### Messages
+
+ - `muffle_messages()` is an internal function that will muffle any messages
+   that originate from the {cli} or {pegboard} packages.
+ - If the {cli} package is not available, messages will have the class of
+   `pbMessage`, which will allow end users/package authors to catch and
+   manipulate any messages that originate from {pegboard}
+
+# pegboard 0.2.7 (unreleased, no user-visible changes)
 
 ## TRANSFORMATION